requirements.txt
decord==0.6.0
fastapi==0.116.1
langchain==0.3.27
langchain_core==0.3.72
langchain_openai==0.3.28
librosa==0.11.0
matplotlib==3.10.3
moviepy==2.2.1
numpy==2.3.2
opencv_contrib_python==4.11.0.86
opencv_python==4.8.1.78
pandas==2.3.1
Pillow==11.3.0
plotly==6.2.0
pydantic==2.11.7
PyJWT==2.10.1
PyJWT==2.10.1
PyMySQL==1.1.0
python-dotenv==1.1.1
reportlab==4.4.2
Requests==2.32.4
scikit_learn==1.7.1
SQLAlchemy==2.0.36
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.53.2
uvicorn==0.35.0
websocket_client==1.8.0
Werkzeug==3.1.3
